requirements.txt
flask==3.0.3
line-bot-sdk==3.12.0
google-api-python-client==2.131.0
google-auth==2.29.0
openai==1.30.5
apscheduler==3.10.4
pytz==2024.1
gunicorn==22.0.0
